Add management RH for picture upload
